Some fixes in code abd in tests

+++ src/main/java/com/training/distance/domain/Route.java
@@ -7,5 +7,5 @@ import java.util.List;
 @Data
 public class Route {
     private List<Distance> distances;
-    private Double fullDistance;
+    private Integer fullDistance;
 }
